+++ src/config.ts
@@ -6,5 +6,5 @@ export default {
   // childMenuWidth: 150, // 自动继承父菜单宽度
   menuItemHeight: 24,
   menuItemDivideLineMargin: 5, // type == 'divide'
-  baseZIndex: 1000, // 基准z-index
+  defaultZIndex: 5000,
 };