+++ src/interface/MenuOption.ts
@@ -1,9 +1,8 @@
+import { PanelOption } from '@/Panel';
 import { BaseAttr } from './common';
 import MenuItemOption from './MenuItemOption';
 
-export default interface MenuOption<Payload> {
-  /** 菜单宽度*/
-  width?: number;
+export default interface MenuOption<Payload> extends PanelOption {
   /** 自定义class */
   class?: BaseAttr<string, Payload>;
   /** 菜单项 */